academic-compass/requirements.txt
streamlit==1.25.0
openai==1.3.7
langchain==0.0.300
faiss-cpu==1.7.4
pandas==1.5.3
PyPDF2==3.0.1
neo4j==5.9.0
networkx==2.8.8
python-dotenv==1.0.0
PyYAML==6.0
pytest==7.2.2
PyMuPDF==1.22.3
sentence-transformers==2.2.2
transformers==4.33.3
torch==2.0.1
keybert==0.7.0
scikit-learn==1.2.2
pytest==7.2.2
rapidfuzz==2.13.0
fuzzywuzzy==0.18.0